feat: split codebase into common and versioned
