+++ .gitignore
@@ -25,7 +25,7 @@ replay_pid*
 
 .gradle
 **/build/
-!src/**/build/
+!core/src/**/build/
 
 # Ignore Gradle GUI config
 gradle-app.setting
