+++ LICENSE_HEADER
@@ -1,0 +1,4 @@
+/*
+ * Copyright Kaylen Dart $YEAR
+ * This project is licensed under the GNU GPLv3 license. See the LICENSE file for more information.
+ */
