+++ common/src/main/java/dog/kaylen/rebrand/config/RebrandModConfig.java
