+++ common/src/main/resources/assets/rebrand/icon.png
