+++ common/src/main/resources/assets/rebrand/lang/es_ar.json
@@ -1,0 +1,10 @@
+{
+    "title.rebrand.config": "Configuración de Rebrand",
+    "option.rebrand.ghostMode": "Habilita Modo Fantasma",
+    "tooltip.rebrand.ghostMode": "Habilita y deshabilita el Modo Fantasma, que obliga al cliente a actuar como un cliente normal durante el protocolo de enlace de inicio de sesión. Esto romperá los servidores multijugador con mods.",
+    "tooltip.rebrand.enable": "Habilita y deshabilita Rebrand para que oculte el nombre de la marca de tu cliente.",
+    "option.rebrand.enable": "Habilita Rebrand",
+    "tooltip.rebrand.brandName": "El nombre de marca (brand) personalizada para enviar a los servidores multijugador al conectarse.",
+    "category.rebrand.general": "General",
+    "option.rebrand.brandName": "Personalizar nombre de marca"
+}
