+++ gradle.properties
@@ -5,12 +5,5 @@ org.gradle.jvmargs=-Xmx1G --add-exports jdk.compiler/com.sun.tools.javac.api=ALL
   --add-exports jdk.compiler/com.sun.tools.javac.tree=ALL-UNNAMED \
   --add-exports jdk.compiler/com.sun.tools.javac.util=ALL-UNNAMED
 
-# Fabric Properties
-# check these on https://modmuss50.me/fabric.html
-minecraft_version=1.20.1
-yarn_mappings=1.20.1+build.9
-loader_version=0.14.21
-# Mod Properties
-mod_version=1.1.1
-maven_group=dog.kaylen
-archives_base_name=rebrand
+java_version = 17
+loader_version=0.15.0
