+++ settings.gradle.kts
@@ -1,0 +1,8 @@
+include("common", "v1_19", "v1_20")
+
+pluginManagement {
+    repositories {
+        maven("https://maven.fabricmc.net/")
+        gradlePluginPortal()
+    }
+}
