--- src/main/resources/assets/rebrand/lang/en_us.json
@@ -1,10 +1,0 @@
-{
-    "title.rebrand.config": "Rebrand Configuration",
-    "category.rebrand.general": "General",
-    "option.rebrand.enable": "Enable Rebrand",
-    "option.rebrand.brandName": "Custom brand name",
-    "option.rebrand.ghostMode": "Enable Ghost Mode",
-    "tooltip.rebrand.enable": "Enables and disables Rebrand from hiding your client brand name.",
-    "tooltip.rebrand.brandName": "The custom brand name to send to multiplayer servers when connecting.",
-    "tooltip.rebrand.ghostMode": "Enables and disables Ghost Mode, which forces the client to act like a vanilla client during the login handshake. This will break multiplayer modded servers."
-}
