--- src/main/resources/rebrand.mixins.json
@@ -1,14 +1,0 @@
-{
-    "required": true,
-    "minVersion": "0.8",
-    "package": "dog.kaylen.rebrand.mixins",
-    "compatibilityLevel": "JAVA_17",
-    "mixins": [],
-    "client": [
-      "ClientBrandRetrieverMixin",
-      "LoginQueryResponseC2SPacketMixin"
-    ],
-    "injectors": {
-        "defaultRequire": 1
-    }
-}
