+++ v1_19/gradle.properties
@@ -1,0 +1,3 @@
+minecraft_version=1.19
+yarn_mappings=1.19+build.4
+loader_version=0.15.0
