+++ v1_19/src/main/resources/rebrand_v1_19.mixins.json
@@ -1,0 +1,14 @@
+{
+    "mixins": [],
+    "minVersion": "0.8",
+    "package": "dog.kaylen.rebrand.mixins",
+    "injectors": {
+        "defaultRequire": 1
+    },
+    "compatibilityLevel": "JAVA_17",
+    "client": [
+        "ClientBrandRetrieverMixin",
+        "LoginQueryResponseC2SPacketMixin"
+    ],
+    "required": true
+}
