+++ v1_20/gradle.properties
@@ -1,0 +1,3 @@
+minecraft_version=1.20
+yarn_mappings=1.20+build.1
+loader_version=0.15.0
